deep_sc/postprocessing: add post_processing_slice_wise
